refactor : 포인트 생성시 상황 고려 책임을 다리로 다시 이동
