+++ src/main/java/controller/LadderController.java
@@ -29,7 +29,6 @@ public class LadderController {
         final Prizes prizes = readWithRetry(this::readPrizes, players);
         final Height height = readWithRetry(this::readHeight, inputView.inputHeight());
 
-//        final Ladder ladder = Ladder.create2(height, PlayerCount.fromPlayers(players), new RandomPointGenerator());
         final Ladder ladder = Ladder.create(height, PlayerCount.fromPlayers(players), new RandomBooleanSupplier());
         outputView.printLadderMap(players, ladder, prizes);
